% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkConnectR-package.R
\docType{package}
\name{NetworkConnectR-package}
\alias{NetworkConnectR}
\alias{NetworkConnectR-package}
\title{NetworkConnectR: A Package to Assess Connectivity in Networks of MPAs}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Rémi Daigle \email{Remi.Daigle@dfo-mpo.gc.ca} (\href{https://orcid.org/0000-0001-8832-4189}{ORCID})

Authors:
\itemize{
  \item Ryan Stanley
  \item Christine Stortini
}

}
\keyword{internal}
